requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
pyjwt==2.3.0
django-mysql==3.9
django-environ==0.8.1
django-cors-headers==3.11.0